Allow specifying a specific version
This works better with tools that want you to pass in the version
number, instead of automatically incrementing it based on what's in
the file. An example of this is in a Makefile where I want to do four
different things with the same version number - it's difficult to
get the version number from this program and then pass it all of the
places it's needed, it's easier to just pass it in and use everywhere.

Updates the successful output to be only the new version number and
a newline. Refactors the code to be more modular and adds a new
SetInFile method for setting a given version number in a file.